Sonar (#370 chunk J): split mission_control/views.py + drop multi-return views
`mission_control.views` was a 1,428-line module; Sonar python:S104 caps
it at 500. Split into a package with focused submodules and a thin
re-exporting facade so existing `from mission_control.views import X`
and `from mission_control import views; views.X` call sites keep
working:

- `mission_control/views/_common.py` (~100) - `_get_user`,
  `_audit_range_lifecycle`, shared constants, and the `_via_pkg` /
  `_pkg()` / `_logger()` late-binding helpers used by submodules.
- `mission_control/views/_pages.py` (~135) - HTML page views
  (`dashboard`, `agents`, `terminal`, `settings`, `help_page`,
  `walkthrough`) and `delete_agent`.
- `mission_control/views/_uploads.py` (~185) - agent presigned-URL
  upload flow (`initiate_upload`, `complete_upload`, `cancel_upload`).
- `mission_control/views/_guacamole.py` (~390) - `guacamole_rdp_url`,
  `api_ngfw_ssh_url`, `guacamole_ssh_url` with shared parse / settings /
  url-generate helpers and a `_SFTP_ROOT_BY_OS` lookup table.
- `mission_control/views/_ranges.py` (~285) - `get_range`,
  `launch_range`, `cancel_range`, `destroy_range`, `pause_range`,
  `resume_range`, `list_agents`, `list_scenarios`. All CMS calls are
  late-bound through the package or via `cms.services` attribute lookup
  so `patch.object(views, "cms_X")` and `patch("cms.services.X")` both
  keep working. The four lifecycle endpoints share a single
  `_dispatch_range_lifecycle` helper parameterised by verb / audit
  action / cms attribute names.
- `mission_control/views/_ngfw.py` (~220) - NGFW HTML pages and JSON
  API; `ngfw_detail` / `ngfw_deprovision` route through
  `_cms_get_ngfw_via_pkg` to honour `patch("mission_control.views.cms_get_ngfw")`.
- `mission_control/views/_credentials.py` (~175) - credential list /
  detail / add pages and JSON API.
- `mission_control/views/_files.py` (~160) - script (`files`,
  `file_upload`, `file_delete`, `api_list_scripts`) views.
- `mission_control/views/__init__.py` (~165) - re-exports every public
  view plus the names tests historically patch at
  `mission_control.views.<name>`: `render`, `logger`,
  `audit_log_from_request`, `AuditLog`, `get_active_range`,
  `get_allowed_extensions`, and the `cms_*` aliases.

Each file stays under SonarCloud's 500-line python:S104 cap.

Drop the 13 multi-return views Sonar flagged under python:S1142 down to
<=3 returns each by extracting guard helpers that raise a private
`_ViewError(JsonResponse)` and orchestrating with a single
try/except/return at the top: `guacamole_rdp_url` (6 -> 2),
`api_ngfw_ssh_url` (7 -> 2), `guacamole_ssh_url` (9 -> 2),
`initiate_upload` (8 -> 2), `launch_range` (7 -> 2), `cancel_range` (4
-> 1), `destroy_range` (4 -> 1), `pause_range` (4 -> 1), `resume_range`
(4 -> 1), `api_ngfw_create` (4 -> 2), `api_ngfw_destroy` (4 -> 2),
`api_credential_create` (6 -> 2), `file_upload` (8 -> 3). The
sftp-root-by-os if-cascade also collapses into the
`_SFTP_ROOT_BY_OS` lookup table to keep `_sftp_root_for_os` under cap.
